pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "github-repo-auditor"
version = "0.17.0"
description = "Automated GitHub portfolio auditor with 12 analysis dimensions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "openpyxl>=3.1.0",
    "radon>=6.0.0",
    "rich>=13.0",
    "anthropic>=0.40.0",
    "fpdf2>=2.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.10"]
config = ["pyyaml>=6.0"]

[project.scripts]
audit = "src.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
